requirements.txt
aiohttp==2.3.10
async-timeout==2.0.0
asyncio==3.4.3
Flask>=0.12.3
Flask-Login>=0.2.7
gunicorn>=19.5.0
Flask-WTF>=0.9.1
watchdog>=0.9.0
lxml>=4.6.2
pymongo==2.8
mongokit-py3>=0.9.1.1
Flask-MongoKit==0.6
bsddb3>=6.2.6
Flask-Mail==0.9.0
Flask-Script==2.0.5
feedparser==5.1.3
sh==1.09
requests>=2.20.0
passlib>=1.7.1
thredds_crawler>=1.5.4
Flask-KVSession==0.6.2
pyyaml>=5.2
rq==1.3.0
rq-dashboard==0.3.4
-e git://github.com/ioos/qartod@12301185bb874fb8193e54822b9ea7da4a34bc9a#egg=ioos_qartod
netCDF4>=1.4.0
cf-units>=2
flask-cors==3.0.2
compliance-checker>=4.3.3
cc-plugin-glider>=2.0.3
flasgger==0.9.5
multidict
ioos-qc
